+++ cmds/start-testnet-api.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+symbol-bootstrap start -p testnet -a api -t target/testnet-api $1