+++ cmds/start-testnet.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 set -e
-symbol-bootstrap start -p testnet -a dual -u current -t target/testnet $1
+symbol-bootstrap start -p testnet -a dual -t target/testnet $1