+++ src/logger/index.ts
@@ -1,5 +1,5 @@
 // created from 'create-ts-index'
 
-export * from './LogType';
 export * from './Logger';
 export * from './LoggerFactory';
+export * from './LogType';