+++ src/service/VotingService.ts
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-import { ConfigParams } from './ConfigService';
-import LoggerFactory from '../logger/LoggerFactory';
-import Logger from '../logger/Logger';
-import { LogType } from '../logger/LogType';
-
-import { BootstrapUtils } from './BootstrapUtils';
-import { ConfigPreset, NodeAccount, NodePreset } from '../model';
 import { join } from 'path';
+import Logger from '../logger/Logger';
+import LoggerFactory from '../logger/LoggerFactory';
+import { LogType } from '../logger/LogType';
+import { ConfigPreset, NodeAccount, NodePreset } from '../model';
+import { BootstrapUtils } from './BootstrapUtils';
+import { ConfigParams } from './ConfigService';
 
 type VotingParams = ConfigParams;
 
@@ -41,7 +40,6 @@ export class VotingService {
             const cmd = [
                 `${presetData.catapultAppFolder}/bin/catapult.tools.votingkey`,
                 `--secret=${nodeAccount.voting.privateKey}`,
-                `--dilution=${presetData.votingKeyDilution}`,
                 `--startEpoch=${presetData.votingKeyStartEpoch}`,
                 `--endEpoch=${presetData.votingKeyEndEpoch}`,
                 `--output=/votingKeys/${privateKeyTreeFileName}`,