+++ test/service/ReportService.test.ts
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-import 'mocha';
-import { BootstrapUtils, ConfigParams, ConfigService, Preset } from '../../src/service';
-import { ReportService } from '../../src/service/ReportService';
 import { expect } from '@oclif/test';
-import { readdirSync } from 'fs';
+import { existsSync } from 'fs';
+import 'mocha';
 import { join } from 'path';
+import { BootstrapUtils, ConfigParams, ConfigService, Preset, ReportService } from '../../src/service';
 
 describe('ReportService', () => {
     const assertReport = async (params: ConfigParams, expectedReportsFolder: string): Promise<void> => {
@@ -27,12 +26,16 @@ describe('ReportService', () => {
 
         const paths = await new ReportService('.', params).run(configResult.presetData);
         const expectedReportFolder = `./test/reports/${expectedReportsFolder}`;
-        expect(paths.length).to.eq(readdirSync(expectedReportFolder).length);
+        await BootstrapUtils.mkdir(expectedReportFolder);
 
         const promises = paths.map(async (reportPath) => {
             expect(reportPath.indexOf(`${params.target}/report`)).to.be.greaterThan(-1);
             const generatedReport = await BootstrapUtils.readTextFile(reportPath);
-            const expectedReport = await BootstrapUtils.readTextFile(join(expectedReportFolder, reportPath.replace(/^.*[\\\/]/, '')));
+            const expectedReportPath = join(expectedReportFolder, reportPath.replace(/^.*[\\\/]/, ''));
+            if (!existsSync(expectedReportPath)) {
+                await BootstrapUtils.writeTextFile(expectedReportPath, generatedReport.trim());
+            }
+            const expectedReport = await BootstrapUtils.readTextFile(expectedReportPath);
             expect(
                 generatedReport.trim(),
                 `Report ${reportPath} doesn't match
@@ -40,7 +43,6 @@ describe('ReportService', () => {
 `,
             ).to.be.eq(expectedReport.trim());
         });
-
         await Promise.all(promises);
     };
 
@@ -90,6 +92,29 @@ describe('ReportService', () => {
         await assertReport(params, 'testnet-peer');
     });
 
+    it('ReportService testnet api', async () => {
+        const target = 'target/ReportService.testnet.api.report';
+        const customPresetObject = {
+            nodes: [
+                {
+                    voting: false,
+                    friendlyName: 'myFriendlyName',
+                },
+            ],
+        };
+        const params = {
+            ...ConfigService.defaultParams,
+            reset: true,
+            preset: Preset.testnet,
+            customPresetObject: customPresetObject,
+            assembly: 'api',
+            target: target,
+            report: true,
+        };
+
+        await assertReport(params, 'testnet-api');
+    });
+
     it('ReportService bootstrap report', async () => {
         const target = 'target/ReportService.bootstrap.voting.report';
         const customPresetObject = {
@@ -112,7 +137,7 @@ describe('ReportService', () => {
         };
         const params = {
             ...ConfigService.defaultParams,
-            reset: false,
+            reset: true,
             preset: Preset.bootstrap,
             customPresetObject: customPresetObject,
             target: target,